pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "reaper-mcp-server"
version = "0.1.0"
description = "MCP server for controlling REAPER DAW with AI agents"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Youssef Hemimy", email = "itsuzef@users.noreply.github.com" },
]
keywords = ["mcp", "reaper", "daw", "audio", "music", "midi", "mastering", "mixing"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Topic :: Multimedia :: Sound/Audio",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
]

dependencies = [
    "mcp>=1.0.0",
    "python-reapy>=0.10.0",
    "numpy>=1.24.0",
    "librosa>=0.10.0",
    "pyloudnorm>=0.1.1",
    "soundfile>=0.12.0",
    "platformdirs>=3.0.0",
]

[project.urls]
Homepage = "https://github.com/bonfire-audio/reaper-mcp"
Repository = "https://github.com/bonfire-audio/reaper-mcp"
Issues = "https://github.com/bonfire-audio/reaper-mcp/issues"

[project.scripts]
reaper-mcp-server = "reaper_mcp.__main__:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/reaper_mcp"]

[tool.ruff]
line-length = 100
target-version = "py310"
